feat: make responsive + strict type
