+++ .gitignore
@@ -21,7 +21,7 @@
 npm-debug.log*
 
 # env files
-# .env
+.env
 
 # vercel
 .vercel
