+++ src/app/page.tsx
@@ -1,34 +1,16 @@
-import Card from "@/components/Card";
+import Card, { CountryType } from "@/components/Card";
 import Nav from "@/components/Nav";
 import { getData } from "@/lib/getData";
 
-interface Article {
-  title: string;
-  author: string;
-  source: string;
-  publishedAt: string;
-  url: string;
-}
-
-interface CountryType {
-  name: "UK" | "US" | "France" | "Australia" | "India";
-  code: string;
-  articles: Article;
-}
-
-interface HomeType {
-  countries: CountryType[];
-}
+// TODO: filter values by country for mobile
 export default async function Home() {
   const countries = await getData();
 
   const CountriesCard = () => {
-    return countries?.map((country: any, index: any) => {
+    return countries?.map((country, index) => {
       const name = country?.name;
       const articles = country?.articles;
-      const cardProps = { name, articles };
-
-      console.log(cardProps);
+      const cardProps = { name, articles } as CountryType;
       return <Card props={cardProps} key={index} />;
     });
   };
@@ -36,7 +18,7 @@ export default async function Home() {
   return (
     <>
       <Nav />
-      <div className="flex flex-row justify-center pt-14">
+      <div className="flex flex-col sm:flex-row justify-center pt-14">
         <CountriesCard />
       </div>
     </>
