+++ src/components/Card.tsx
@@ -1,9 +1,28 @@
-export default function Card({ props }: any) {
+interface Article {
+  title: string;
+  author: string;
+  source: {
+    name: string;
+  };
+  publishedAt: string;
+  url: string;
+}
+
+export interface CountryType {
+  name: string;
+  articles: Article[];
+}
+
+interface CardProps {
+  props: CountryType;
+}
+
+export default function Card({ props }: CardProps): JSX.Element {
   const DisplayHeadlines = () => {
     const hasArticles = props?.articles;
     let list;
     if (hasArticles) {
-      list = props?.articles.map((article: any, index: any) => {
+      list = props?.articles.map((article: Article, index: number) => {
         return (
           <li className="pb-6 pt-4" key={index}>
             <p>
@@ -35,10 +54,10 @@ export default function Card({ props }: any) {
 
   return (
     <div className="min-w-lg rounded overflow-hidden shadow-lg px-6">
-      <h1 className="flex items-center text-4xl font-extrabold text-red-900">
+      <h1 className="lg:text-4xl items-center text-2xl font-extrabold text-red-900">
         {props?.name}
       </h1>
-      <span className="bg-blue-100 text-2xl font-semibold me-2 px-2.5 py-0.5 rounded dark:bg-red-200 dark:text-black ms-2">
+      <span className="bg-blue-100 lg:text-2xl text-xl font-semibold me-2 px-2.5 py-0.5 rounded dark:bg-red-200 dark:text-black ms-2">
         Top 5 Headlines
       </span>
       <DisplayHeadlines />
